Source code for the 1.3.5 PyPI release
This is the closest we have to the commit that was released to PyPI as
version 1.3.5 on 2017-01-30, with filename pypyodbc-1.3.5.2.zip.

A diff between `python setup.py sdist` and the file on PyPI is
committed. The differences only concern PKG-INFO metadata.

Unfortunately, setup.py claims the version is 1.3.4, which caused
warnings and errors from pip down the road.